+++ engine-io/default.nix
@@ -1,7 +1,7 @@
 { mkDerivation, aeson, async, attoparsec, base, base64-bytestring
 , bytestring, errors, free, monad-loops, mwc-random, stdenv, stm
 , stm-delay, text, transformers, unordered-containers, vector
-, websockets
+, websockets, zlib
 }:
 mkDerivation {
   pname = "engine-io";
@@ -10,7 +10,7 @@ mkDerivation {
   libraryHaskellDepends = [
     aeson async attoparsec base base64-bytestring bytestring errors
     free monad-loops mwc-random stm stm-delay text transformers
-    unordered-containers vector websockets
+    unordered-containers vector websockets zlib
   ];
   homepage = "http://github.com/ocharles/engine.io";
   description = "A Haskell implementation of Engine.IO";